refactor: DTO와 내부 모델 분리를 위해 CompanyInfo, HubInfo를 model 패키지로 이동
