+++ src/main/java/org/gupang/deliveryservice/application/service/CompanyService.java
@@ -1,6 +1,6 @@
 package org.gupang.deliveryservice.application.service;
 
-import org.gupang.deliveryservice.application.dto.CompanyInfo;
+import org.gupang.deliveryservice.application.model.CompanyInfo;
 
 import java.util.UUID;
