+++ src/main/java/org/gupang/deliveryservice/application/service/DeliveryService.java
@@ -4,9 +4,9 @@ import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import org.gupang.common.exception.CustomException;
 import org.gupang.common.exception.ErrorCode;
-import org.gupang.deliveryservice.application.dto.CompanyInfo;
+import org.gupang.deliveryservice.application.model.CompanyInfo;
 import org.gupang.deliveryservice.application.dto.CreateDeliveryCommand;
-import org.gupang.deliveryservice.application.dto.HubInfo;
+import org.gupang.deliveryservice.application.model.HubInfo;
 import org.gupang.deliveryservice.domain.entity.Delivery;
 import org.gupang.deliveryservice.domain.repository.DeliveryRepository;
 import org.springframework.stereotype.Service;
