+++ src/main/java/org/gupang/deliveryservice/application/service/HubService.java
@@ -1,6 +1,6 @@
 package org.gupang.deliveryservice.application.service;
 
-import org.gupang.deliveryservice.application.dto.HubInfo;
+import org.gupang.deliveryservice.application.model.HubInfo;
 
 import java.util.List;
 import java.util.UUID;
